Standalone apply router
